wip - seperate hacky toolchain

+++ src/firmware_new/tools/debug_firmware_on_arm_board.sh
@@ -11,19 +11,29 @@ if [ $NUM_ARGS -ne 2 ]; then
     exit 1
 fi
 
+OPENOCD_CFG_FILE=$2
+ELF_FILE=$1
+
+# Make sure target we're loading actually exists
+if [ ! -f $ELF_FILE ]; then
+    echo "File given does not exist: '$ELF_FILE'"
+    exit 1
+fi
+
+
 # Stop openocd if it's already running
 pkill openocd
 KILLED_OPENOCD=$?
 
 # If we killed openocd, wait for a second for the port to be free
-if [ $KILLED_OPENOCD ]; then
+if [ $KILLED_OPENOCD = 0 ]; then
     echo "Had to kill instance of openocd, waiting for a second for it to stop"
     sleep 2
 fi
 
 # Start openocd as a child process, and redirect all its output to /dev/null.
 # If we don't redirect the output it will overwrite the gdb window
-OPENOCD_RUN_CMD="openocd -f $2"
+OPENOCD_RUN_CMD="openocd -f $OPENOCD_CFG_FILE"
 eval "$OPENOCD_RUN_CMD > /dev/null 2>&1 &"
 
 # Check that openocd is still running after a second. If it's not, that 
@@ -31,7 +41,7 @@ eval "$OPENOCD_RUN_CMD > /dev/null 2>&1 &"
 sleep 1
 pgrep openocd 
 OPENOCD_RUNNING=$?
-if [ ! $OPENOCD_RUNNING ]; then
+if [ $OPENOCD_RUNNING -ne 0 ]; then
     echo "There was an issue running openocd, the command that failed was: '$OPENOCD_RUN_CMD'"
     exit 1
 fi
@@ -44,7 +54,7 @@ arm-none-eabi-gdb \
     -ex "directory $BAZEL_ROOT_DIR" \
     -ex "target extended-remote :3333" \
     -ex "load" \
-    -tui $1
+    -tui $ELF_FILE
 
 # When gdb is finished, cleanup by stopping openocd
 pkill openocd
